requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
pandas>=1.2.0

# Computer vision
opencv-python>=4.5.1
scikit-image>=0.18.1
albumentations>=1.0.0
monai>=0.7.0

# Deep learning frameworks
ultralytics>=8.0.0
timm>=0.4.12

# Medical imaging
pydicom>=2.1.2
SimpleITK>=2.1.0

# Visualization
matplotlib>=3.3.2
seaborn>=0.11.1
plotly>=4.14.3

# Utilities
tqdm>=4.56.0
PyYAML>=5.4.1
scikit-learn>=0.24.1
pillow>=8.2.0

# For development
black>=21.5b2
isort>=5.8.0
flake8>=3.9.1
mypy>=0.812
pytest>=6.2.3
